requirement.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.6
sqlparse==0.2.4
tzdata==2022.6
